% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_read_and_write.R
\name{doc_ggsave}
\alias{doc_ggsave}
\title{doc_ggsave}
\usage{
doc_ggsave(
  plot,
  milestone,
  filename,
  device = "png",
  width = 10,
  height = 6,
  units = "in",
  ...
)
}
\arguments{
\item{plot}{}

\item{milestone}{}

\item{filename}{}

\item{device}{}

\item{width}{}

\item{height}{}

\item{units}{}

\item{...}{}
}
\description{
doc_ggsave
}
