% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_edges_and_vertices.R
\name{doc_multi_vertex}
\alias{doc_multi_vertex}
\title{doc_multi_vertex}
\usage{
doc_multi_vertex(name, milestone, folder, files = NULL, selector = NULL)
}
\arguments{
\item{name}{}

\item{milestone}{}

\item{folder}{}

\item{files}{}

\item{selector}{}
}
\description{
doc_multi_vertex
}
