% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_plot.R
\name{doc_plot}
\alias{doc_plot}
\title{doc_plot}
\usage{
doc_plot(list_data_vertices = NULL, list_data_edges = NULL)
}
\arguments{
\item{list_data_vertices}{}

\item{list_data_edges}{}
}
\description{
doc_plot
}
