% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_reactable.R
\name{doc_reactable}
\alias{doc_reactable}
\title{doc_reactable}
\usage{
doc_reactable(
  table,
  fontsize = 14,
  headersize = 16,
  resizable = FALSE,
  groupBy = NULL,
  total_width = 780,
  percent_vector = NULL
)
}
\arguments{
\item{table}{}

\item{fontsize}{}

\item{headersize}{}

\item{resizable}{}

\item{groupBy}{}

\item{total_width}{}

\item{percent_vector}{}
}
\description{
doc_reactable
}
