% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_read_and_write.R
\name{doc_read_data}
\alias{doc_read_data}
\title{doc_read_data}
\usage{
doc_read_data(
  file,
  milestone,
  sep = ",",
  encoding = "UTF-8",
  sheet = NULL,
  pagesize = 1e+05,
  id = NA_character_,
  ...
)
}
\arguments{
\item{file}{file to read}

\item{milestone}{project step or working package}

\item{sep}{optional seperator for csv files}

\item{encoding}{encoding}

\item{sheet}{sheet}

\item{pagesize}{only relevant for json files}

\item{id}{optional id to count unique values}

\item{...}{arguments to pass to the reading function}
}
\description{
doc_read_data
}
