% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_setup.R
\name{doc_save}
\alias{doc_save}
\title{Save current state of the documentation}
\usage{
doc_save(top_directory = NULL)
}
\arguments{
\item{top_directory}{folder that contains the documentation folder, if null assumes that the documentation folder is in the current working directory}
}
\description{
Save current state of the documentation
}
