% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_sheet.R
\name{doc_sheet}
\alias{doc_sheet}
\title{Create a Report Object that can interactively create a rmd html document}
\usage{
doc_sheet(
  title,
  author = "",
  theme = "default",
  bs_version = 5,
  filemanager = "browseURL"
)
}
\arguments{
\item{title}{}

\item{author}{}

\item{theme}{a bootswatch theme}

\item{bs_version}{}

\item{filemanager}{}
}
\description{
Create a Report Object that can interactively create a rmd html document
}
