% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_setup.R
\name{doc_start}
\alias{doc_start}
\title{Creates documentation folder in current working directory and sets up files}
\usage{
doc_start()
}
\description{
Creates documentation folder in current working directory and sets up files
}
