% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_read_and_write.R
\name{doc_write_data}
\alias{doc_write_data}
\title{doc_write_data}
\usage{
doc_write_data(data, file, milestone, id = NA_character_, ...)
}
\arguments{
\item{data}{}

\item{file}{}

\item{milestone}{}

\item{id}{}

\item{...}{}

\item{sheetName}{}
}
\description{
doc_write_data
}
