% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_edges_and_vertices.R
\name{new_data_edge}
\alias{new_data_edge}
\title{new_data_edge}
\usage{
new_data_edge(
  from,
  to,
  comments = "none",
  path = rstudioapi::getSourceEditorContext()$path
)
}
\arguments{
\item{from}{}

\item{to}{}

\item{comments}{}

\item{path}{}
}
\description{
new_data_edge
}
