% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_edges_and_vertices.R
\name{new_data_vertex}
\alias{new_data_vertex}
\title{new_data_vertex}
\usage{
new_data_vertex(
  df = NA_character_,
  filename = NA_character_,
  dirname = NA_character_,
  id = NA_character_,
  Meilenstein = NA_character_
)
}
\arguments{
\item{df}{}

\item{filename}{}

\item{dirname}{}

\item{id}{}

\item{Meilenstein}{}
}
\description{
new_data_vertex
}
