% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_sheet.R
\name{Report}
\alias{Report}
\title{R6 Class representing a HTML Report that can modify and render itself}
\description{
R6 Class representing a HTML Report that can modify and render itself

R6 Class representing a HTML Report that can modify and render itself
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Report-new}{\code{Report$new()}}
\item \href{#method-Report-add_object}{\code{Report$add_object()}}
\item \href{#method-Report-add_text}{\code{Report$add_text()}}
\item \href{#method-Report-add_df_info}{\code{Report$add_df_info()}}
\item \href{#method-Report-open}{\code{Report$open()}}
\item \href{#method-Report-reload}{\code{Report$reload()}}
\item \href{#method-Report-render}{\code{Report$render()}}
\item \href{#method-Report-save}{\code{Report$save()}}
\item \href{#method-Report-clone}{\code{Report$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Report-new"></a>}}
\if{latex}{\out{\hypertarget{method-Report-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$new(
  title,
  author = "",
  theme = "default",
  bs_version = 5,
  filemanager = "browseURL"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Report-add_object"></a>}}
\if{latex}{\out{\hypertarget{method-Report-add_object}{}}}
\subsection{Method \code{add_object()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$add_object(
  object,
  name = NULL,
  header = NULL,
  footer = NULL,
  width = 8,
  height = 4,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Report-add_text"></a>}}
\if{latex}{\out{\hypertarget{method-Report-add_text}{}}}
\subsection{Method \code{add_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$add_text(text, level = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Report-add_df_info"></a>}}
\if{latex}{\out{\hypertarget{method-Report-add_df_info}{}}}
\subsection{Method \code{add_df_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$add_df_info(data, id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Report-open"></a>}}
\if{latex}{\out{\hypertarget{method-Report-open}{}}}
\subsection{Method \code{open()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$open()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Report-reload"></a>}}
\if{latex}{\out{\hypertarget{method-Report-reload}{}}}
\subsection{Method \code{reload()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$reload()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Report-render"></a>}}
\if{latex}{\out{\hypertarget{method-Report-render}{}}}
\subsection{Method \code{render()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$render()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Report-save"></a>}}
\if{latex}{\out{\hypertarget{method-Report-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$save(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Report-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Report-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
