% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{search_files}
\alias{search_files}
\title{search_files}
\usage{
search_files(search_pattern, file_pattern = "R$", path = ".", recursive = T)
}
\arguments{
\item{search_pattern}{pattern to search for within the files}

\item{file_pattern}{file names to look for}

\item{path}{where to search}

\item{recursive}{search path recursively or not}
}
\description{
search_files
}
